requirements.txt
langchain==0.3.14
langchain-core==0.3.29
langchain-community==0.3.14
langchain-groq==0.2.3
langchain-text-splitters==0.3.5
langgraph==0.2.73
chromadb==0.6.3
tavily-python==0.5.0
streamlit==1.41.1
python-dotenv==1.0.1
tiktoken==0.8.0
sentence-transformers==3.3.1
huggingface-hub==0.27.1